Implements single character cyphering
Using the basic configuration of the Enigma machine as a guide, we'll
use three rotors to implement a simple substitution (caesar) cypher.

The input character is changed to an integer, mutated three times with
the rotor, and translated back to a letter from a number.

We're using the array method `at` here instead of what we might usually
do with using the `[]` access of an array index. This is to highlight
the structure of the messages being passed. It's much easier to think of
`.at` being a message passed from the Enigma's cypher method to an array
(the rotor), than `ROTOR_I[index]`.

>> Even if the `[]` accessor is also a method call equivalent to
`ROTOR_I.[](index)`. 😜