Extracts index shifting to private method
While the tests are green, we'll do some refactoring.

Keeping in mind the DRY principle, or the single responsibility
principle, we keep our attention on the rotor's `.at` calls.

Are those steps in our algorythm conceptually the same? Meaning: if we
wanted to change the way that our machine scrambles an integer into
another, would all three `.at` calls change?

Yes!

So instead of having to change the `.at` in three different places, we
can extract a method, and point to a single source of truth for the
shifting action. There may stil be repeated text in the file.. but our
code is now more DRY because we've created that source of truth and
gathered together all the functionality that would change **for the same
reason**.