Extracts Lampboard and Keyboard classes
While we're thinking about DRY, it's easy to move into thinking about
the Single Responsibility Principle.

> A module should be responsible to only one actor

Actor in this case means something (whether a business requirement or
another module in the codebase) that would require the module to change.

Let's think about a use case where we would need to change the type of
input we accept, an image of a letter instead of a string character for
instance.

Previously the Enigma class would have to change to accommodate this new
requirement. The same class would also need to respond to a change in
how the encryption is handled. "The input converting to index" and "the
encryption method" are two different actors our code is responsible to.

We're also thinking about the Enigma device itself. These principles
don't exist purely to satisfy themselves, but within the context of what
we're trying to accomplish. This leads to the idea of replicating some
of the concepts of the device itself, a keyboard and a lampboard.

The keyboard on the physical device transforms a keypress into an
electrical current, our keyboard transforms string input into an
integer. The lampboard takes that current and transforms it into light,
our lampboard returns an integer to character form.