Adds test for advancing cypher
Before we repeat the same simple extraction for a Rotor class similar to
the Keyboard and Lampboard, we'll introduce a new test that replicates
some of the behavior that elevated the Enigma machine above a simple
substitution cypher.

When the same character was entered into the machine twice in a row, two
different characters would be returned in sequence. The cypher key
changed after every keypress in a repeatable way if the beginning
settings were the same.

Before stepping to the next commit, I encourage you to take the Enigma
code as we've created it so far and try for yourself. Think about how we
did the simple class extraction of a Rotor using DRY and Single
Responsibility first. Write tests for the next Rotor class and then add
the complexity of the step motion.

If you're feeling saucy, uncomment the last test and add the ability for
the machine to encypher a phrase.

When you've gotten the new test(s) to pass, take a look thruogh the next
commits to see how I accomplished it.