Extracts Rotor class
We've taken the first step now, extracted a simple Rotor class that does
the behavior that already existed in the Enigma class internals. This
allows us to delete the private `Enigma#translate` method altogether.

How does this change our understanding of the responsibility of the
Enigma class?

It's no longer handling any of the conversions, index, character or
otherwise. The Enigma class is now operating mainly as an orchestrator.
The responsibility is to pass messages around to its collaborators, but
each individual step is accomplished by a specialized module.

This matches the idea of a physical device too. The "Enigma Machine" is
an abstraction we use to refer to a combination of parts that make up a
whole. At each step further into the system we recognize smaller and
more specialized components.

Here we're choosing to operate at the level of abstraction that is
larger parts of the entire device; rotors, keyboards, etc. We have code
abstractions that don't require us to think at the level of wires and
switches, but those components might be represented by the code that
makes up the ruby `Array#at` method.