Adds position to Rotor class
In order for the rotor to return different characters along the length
of its key, it is going to have to store how many times it has been used
(or the equivalent) to know how much to step forward.

Now our rotors will move forward one index in their keys after each call
to `#translate`.

The combination of related behavior and state is the core idea of an
object, or a class. Now not only do we have some state that will be
given to the Rotor from a collaborator (the key) but also some state
that is internal only to it's own operation.

We're not going to expose the value of the position to the Rotor
objects' collaborators, because its not necessary. We're using another
principle of object oriented design to **encapsulate** the position
state within the rotor.

We are allowing collaborators (the Enigma class) to send the message
`translate` to a rotor, and trusting it will receive a new integer. How
it translates it is not of the Enigma object's concern, and we've
successfully kept the Enigma responsible to only a single actor.

If the Enigma class was to contain code that used the implementation of
the stepping mechanism, we're exposing it to the need to change if that
process changes. That sharing of an actor is **coupling**. Two objects
are coupled when changes in one require changes in another.
Encapsulation is one way to avoid unneccessary coupling.