Encypher phrases
We've uncommented the test for being able to encypher phrases and
updated our Enigma class to iterate over a string. But when we run this
test...

> # Running:
> ......F....
> Finished in 0.000587s, 18739.3512 runs/s, 32367.9702 assertions/s.
>
> 1) Failure:Enigma::Test#test_encyphers_a_phrase [enigma_test.rb:28]:
>      Expected: "IPNNYTYENR"
>      Actual: "IMIZTPSRHJ"
>
> 11 runs, 19 assertions, 1 failures, 0 errors, 0 skips

Our machine is not currently operating the way we would expect. To
understand the issue we'll have to think about how the physical device
works.

After a keypress and translation, the first rotor rotates, but only
moves the second rotor if it's made a full rotation. Right now our
device always steps all the rotors. This gives us a new key each time,
but there are only 26 different possible keys, the real device had
17,576.