Adds advancing rotor mechanism
We've added functionality to the rotor objects that will let the outside
world know when they've ticked over and the next cog should advance. Its
this that enables the 17,000+ different cypher keys that change with
every letter, the mechanism that made Enigma so formidable.

But in creating this featue, we've increased the coupling between the
rotors and the enigma class. The `Enigma#cypher` method might not be
able to directly change the position state of the rotor, but it's
involved in the process maybe more than it needs to be.

Coupling isn't always bad! Any application is going to have coupling
between collaborators, if no modules ever talked to each other, it'd be
a fairly useless program. What we should focus on is a minimal &
maintainable amount of coupling.

There's opportunity here to use OO Design patterns to decrease the
coupling between these classes and make this codebase more friendly to
future changes.