Extracts Spindle Class
We've once again decoupled the Enigma from performing the specific steps
that make up each action in the encyphering process. By sending a
message individually to number of collaborating modules, we've protected
the Enigma class from the necessity of changing any of its own code if
the implementation of its component parts change.