Abstracts over dependencies
By removing even explicit references to the classes that Enigma uses to
function, instead passing them in as arguments, we've abstracted away
any dependency to a specific class or object. Enigma and Spindle are
similarly innoculated against changes in their dependencies because they
rely only on an abstract interface.

This is sometimes referred to as "Dependency Inversion" because the
loose coupling between higher-level modules and their utility modules
**inverts** the assumed relationship of system design.

Listen... I didn't name the thing, ok?

Another OO principle is on display here as well, the open/closed
principle. The Enigma class is "open for extension but closed for
modification," meaning that we can add functionality to the Enigma's
behavior without having to change the code in the Enigma class by
passing it a different collaborator.

As long as the transformer module we send in on initialization responds
to the `call` message, and returns an index that can be consumed by the
output module.. it can be anything and act in any new way.

Enigma is ready for sweeping changes in its functionality as
requirements change, as new features are built, making extension of this
class painless in the future.