Adds test for reflector and decryption
The one missing piece to our emulator replicating a functional Enigma
machine is decryption. Given the same starting settings as the machine
that encyphered a string of characters, the Enigma machine will reverse
the process and DEcrypt the phrase.

We've added new tests in the enigma_test.rb file which will test for
this ability. Make use of the "open for extension" property of our
Enigma emulator and build a `ReflectorSpindle` class which can
accomplish this task and make the tests pass.

Go ahead, the next commit will still be here when you get back.